🎯 Pre-Refactor Commit: Unified Backtest System + Git Traceability
MAJOR ARCHITECTURAL ACHIEVEMENT:
✅ Complete Unified Strategy Model Backtest System
✅ Git Commit Hash Tracking & Traceability System
✅ Comprehensive Documentation Framework

## 🧬 Unified Strategy Model Architecture COMPLETE
- Strategy models now act as 'replaceable units of knowledge'
- IStrategyModel interface for signal generation only
- StrategyModelFactory for dynamic model loading from YAML
- ODTE.Execution handles ALL fills (strategy-agnostic)
- Complete separation: Strategies generate signals, Execution handles fills

## 📋 Git Commit Hash Traceability System COMPLETE
- Pre-backtest git state verification with warnings
- Full git commit hash tracking in all reports
- Working tree status detection (clean vs uncommitted changes)
- Legitimacy flags: VALID ✅ (clean) vs FLAGGED ⚠️ (uncommitted)
- Complete audit trail: git commit → config → execution → results

## 📚 Comprehensive Documentation COMPLETE
- ODTE.Strategy/BACKTEST_EXECUTION_GUIDE.md (Strategy development)
- ODTE.Backtest/BACKTEST_EXECUTION_GUIDE.md (Unified execution)
- ODTE.Historical/BACKTEST_EXECUTION_GUIDE.md (Data providers)
- ODTE.Execution/BACKTEST_EXECUTION_GUIDE.md (Execution engine)
- Mandatory git commit requirements documented across all projects

## 🎯 SPX30DTE Implementation
- SPX30DTEStrategyModel implementing IStrategyModel
- Multi-component strategy: BWB + Probes + VIX Hedging + RevFib
- Genetic algorithm traceability: 16_mutation_tournament
- Configuration-driven execution via SPX30DTE_v1.0_config.yaml

## 🔧 Technical Implementation
- Enhanced Program.cs with git information retrieval
- GitInformation class for commit hash tracking
- Updated report generation with git traceability section
- Registry updates with git commit documentation
- Working tree status warnings for reproducibility

TESTING VERIFIED:
- Git information capture working (commit: f3edd7fba7252d943808779329964975e3a9203f)
- SPX30DTE backtest execution through unified system
- Automatic report generation with git traceability
- Registry updates with legitimacy validation

EXECUTION COMMAND:
cd ODTE.Backtest && dotnet run '../ODTE.Configurations/Models/[Model]_v[Version]_config.yaml'

This commit establishes the foundation for the unified backtest system
with complete git traceability. Ready for deep refactor with fallback
reference to this working implementation.

Status: ✅ PRODUCTION READY - Unified Architecture Complete

+++ ODTE.Strategy/SPX30DTE/Mutations/SimpleTournamentDemo/SimpleTournamentDemo.cs
@@ -1,7 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-
 namespace ODTE.Strategy.SPX30DTE.Mutations
 {
     // Simplified demonstration of the 16-mutation tournament concept
@@ -18,7 +14,7 @@ namespace ODTE.Strategy.SPX30DTE.Mutations
 
             Console.WriteLine("🧬 GENERATED 16 STRATEGIC MUTATIONS");
             Console.WriteLine("━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━");
-            
+
             for (int i = 0; i < mutations.Count; i++)
             {
                 var mutation = mutations[i];
@@ -44,7 +40,7 @@ namespace ODTE.Strategy.SPX30DTE.Mutations
             Console.WriteLine();
             Console.WriteLine("📊 THEORETICAL SQLITE LEDGER GENERATION");
             Console.WriteLine("━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━");
-            
+
             foreach (var result in tournamentResults.Take(4))
             {
                 Console.WriteLine($"🗄️  {result.MutationName}_Ledger_20050101_20250101.db");
@@ -134,7 +130,7 @@ namespace ODTE.Strategy.SPX30DTE.Mutations
                 var preservationScore = (finalValue - 100000m) / 400000m * 25; // 25% weight, capital growth
                 var winRateScore = winRate * 20; // 20% weight
 
-                var overallScore = (decimal)(cagrScore + riskScore + preservationScore + (double)winRateScore);
+                var overallScore = cagrScore + riskScore + preservationScore + winRateScore;
 
                 results.Add(new TournamentResult
                 {
@@ -147,7 +143,7 @@ namespace ODTE.Strategy.SPX30DTE.Mutations
                     TotalTrades = totalTrades,
                     FinalValue = finalValue,
                     ProfitFactor = 1.4m + (decimal)(random.NextDouble() * 1.1), // 1.4-2.5
-                    SharpeRatio = 1.2m + (decimal)(random.NextDouble() * 0.8m), // 1.2-2.0
+                    SharpeRatio = 1.2m + (decimal)(random.NextDouble() * 0.8), // 1.2-2.0
                     OverallScore = overallScore
                 });
             }
@@ -166,7 +162,7 @@ namespace ODTE.Strategy.SPX30DTE.Mutations
                 var trophy = i switch
                 {
                     0 => "🥇",
-                    1 => "🥈", 
+                    1 => "🥈",
                     2 => "🥉",
                     _ when i < 8 => "🏅",
                     _ => "  "
@@ -184,7 +180,7 @@ namespace ODTE.Strategy.SPX30DTE.Mutations
             {
                 var rank = topResults.IndexOf(result) + 1;
                 var medal = rank switch { 1 => "🥇", 2 => "🥈", 3 => "🥉", _ => "🏅" };
-                
+
                 Console.WriteLine($"{medal} #{rank}: {result.MutationName}");
                 Console.WriteLine($"   💰 20-Year Growth: ${result.FinalValue:N0} (CAGR: {result.CAGR:P1})");
                 Console.WriteLine($"   📉 Max Drawdown: {result.MaxDrawdown:P1} | 🎯 Win Rate: {result.WinRate:P1}");
@@ -201,7 +197,7 @@ namespace ODTE.Strategy.SPX30DTE.Mutations
             var tradeDataSize = totalTrades * 0.5m; // 500 bytes per trade in KB
             var legDataSize = totalTrades * 3 * 0.2m; // Average 3 legs per trade, 200 bytes each
             var dailyPnLSize = 20 * 365 * 0.1m; // 20 years of daily P&L, 100 bytes per day
-            
+
             return (tradeDataSize + legDataSize + dailyPnLSize) / 1024; // Convert KB to MB
         }
